requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.0.3
packaging==26.0
psycopg==3.3.2
psycopg-binary==3.3.2
PyJWT==2.11.0
sqlparse==0.5.5
typing_extensions==4.15.0
